requirements.txt
boto3==1.9.50
botocore==1.12.50
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.3
djangorestframework==3.9.0
docutils==0.14
idna==2.7
jmespath==0.9.3
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1

